requirements.txt
streamlit==1.33.0
torch==2.2.0
torchvision==0.17.0
pillow==10.2.0
numpy==1.26.4
pyarrow==15.0.0